+++ src/Patient/Onboarding/index.js
@@ -1,24 +1,22 @@
-import React, { useState } from 'react'
+import React, { useState } from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import OverTopBar from '../Navigation/OverTopBar';
 import useStores from '../../Basics/UseStores';
-import { ReactComponent as DoctorIcon } from '../../Basics/Icons/DoctorGroup.svg';
 import SimpleButton from '../../Basics/SimpleButton';
 import { useTranslation } from 'react-i18next';
-import CheckIcon from '@material-ui/icons/Check';
-import XIcon from '@material-ui/icons/Clear';
-import Styles from '../../Basics/Styles';
 import MobileStepper from '@material-ui/core/MobileStepper';
-import Gender from './Gender'
+import Gender from './Gender';
 import Age from './Age';
-import Notification from './Notification'
-import ContactTracing from './ContactTracing'
-import End from './End'
-import Password from './Password'
-import { observer } from 'mobx-react'
+import Notification from './Notification';
+import ContactTracing from './ContactTracing';
+import End from './End';
+import Password from './Password';
+import { observer } from 'mobx-react';
 import CircularProgress from '@material-ui/core/CircularProgress';
-
-import PushPermissionsNotice from "./PushPermissionNotice"
+import Landing from './Landing';
+import PushPermissionsNotice from './PushPermissionNotice';
+import AssistantFAQ from './AssistantFAQ';
+import Styles from '../../Basics/Styles';
 
 const useStyles = makeStyles({
     body: {
@@ -37,44 +35,9 @@ const useStyles = makeStyles({
         left: 0,
         zIndex: 10,
     },
-    landing: {
-        display: "flex",
-        flexDirection: "column",
-        alignItems: "center",
-        padding: "1em",
-        textAlign: "center"
-    },
     button: {
         position: "fixed",
         bottom: "2em"
-    },
-    faq: {
-        height: "60vh",
-        ...Styles.flexColumn,
-        justifyContent: "space-evenly",
-        padding: "1em",
-        "& > h1": { fontSize: "1.5em" },
-        "& > h2": { fontSize: "1.25em", margin: 0, padding: 0 },
-        "& > ul > li": {
-            listStyle: "none",
-            display: "flex",
-            alignItems: "center",
-            "& > svg": {
-                fontSize: ".8em",
-                marginRight: ".5em"
-            }
-        }
-    },
-    green: {
-        color: "green"
-    },
-    red: {
-        color: "red"
-    },
-    bottom: {
-        borderTop: "1px solid lightgray",
-        padding: "1em",
-        marginTop: "auto"
     },
     stepper: {
         backgroundColor: "white",
@@ -94,49 +57,7 @@ const useStyles = makeStyles({
     }
 })
 
-const Landing = () => {
-    const { patientStore } = useStores();
-    const classes = useStyles();
-    const { t, i18n } = useTranslation('translation');
-
-    return (
-        <div className={classes.landing}>
-            <DoctorIcon />
-            <h2>{t('patient.onboarding.landing.welcome')} {patientStore.givenName}</h2>
-            <p>{t('patient.onboarding.landing.message')}</p>
-        </div>
-    )
-}
-
-const CoordinatorFAQ = () => {
-    const classes = useStyles();
-    const { t } = useTranslation('translation');
-
-    return (
-        <div className={classes.faq}>
-            <h1>{t('patient.onboarding.coordinator.title')}</h1>
-            <h2>{t('patient.onboarding.coordinator.will.title')}</h2>
-            <ul>
-                {t('patient.onboarding.coordinator.will.items', { returnObjects: true }).map(each => {
-                    return <li> <CheckIcon className={classes.green} /> {each}</li>
-                })}
-            </ul>
-
-            <h2>{t('patient.onboarding.coordinator.wont.title')}</h2>
-            <ul>
-                {t('patient.onboarding.coordinator.wont.items', { returnObjects: true }).map(each => {
-                    return <li><XIcon className={classes.red} /> {each}</li>
-                })}
-            </ul>
-
-            <div className={classes.bottom}>
-                <p>{t('patient.onboarding.coordinator.outro')}</p>
-            </div>
-        </div>
-    )
-}
-
-const Tabs = [<Landing />, <CoordinatorFAQ />, <Password overrideNext />, <Gender />, <Age />, <PushPermissionsNotice overrideNext />, <Notification />, <ContactTracing />, <End overrideNext />]
+const Tabs = [<Landing />, <AssistantFAQ />, <Password overrideNext />, <Gender />, <Age />, <PushPermissionsNotice overrideNext />, <Notification />, <ContactTracing />, <End overrideNext />]
 
 const Onboarding = observer(() => {
 
@@ -144,15 +65,9 @@ const Onboarding = observer(() => {
     const { patientStore, activationStore, patientUIStore } = useStores();
     const { t } = useTranslation('translation');
 
-    const index = patientUIStore.reportStep
+    const index = patientUIStore.reportStep;
 
-    const handleNext = () => {
-        if (patientUIStore.reportStep === Tabs.length - 1) {
-            //activationStore.submitActivation();
-        } else {
-            patientUIStore.updateOnboardingStep(index + 1)
-        }
-    }
+    const handleNext = () => {patientUIStore.updateOnboardingStep(index + 1)}
     const handleBack = () => { index < 1 ? patientStore.logout() : patientUIStore.updateOnboardingStep(index - 1) }
 
     return (
